The most significant changes involve the addition of a new method `GenerateCacheWithUrls()` in the `GenerateCacheURL` class, which collects song details, validates URLs, downloads assets, and generates a cache. The method for writing the serialized JSON to a file has been updated, with changes to the file name and path. The `ConvertUbiArtToUnity` class, `KaraokeClip` class, `Structure` and `Signature` classes, and `PInvoke` class have all been updated with changes to property types and method attributes. The `Program` class has been updated with changes to the options in the `Question.Ask` method and the switch-case statements in the `GenerateCache` method. The `Print.PrintCache` method call has been removed from the `GenerateCache` method.
List of changes:

1. Addition of `GenerateCacheWithUrls()` method in `GenerateCacheURL` class, which collects song details, validates URLs, downloads assets, and generates a cache. The method for writing the serialized JSON to a file has been updated, with changes to the file name and path. The variable `jsonCache0` has been renamed to `cachingStatus`. The code for printing the serialized JSON to the console has been removed.
2. Modification of `ConvertUbiArtToUnity` class in `ConvertUbiArtToUnity.cs` file. The `SemitoneTolerance` property is now cast to an integer before being assigned to `karaokeClip["SemitoneTolerance"]`. The `preloadSize` property of `second` is now set to the count of `assetBundleArray.Children`.
3. Update of `KaraokeClip` class in `KaraokeTape.cs` file. The `SemitoneTolerance` property's type has been changed from `int` to `float`.
4. Update of `Structure` and `Signature` classes in `MusicTrack.cs` file. The `previewEntry`, `previewLoopStart`, and `marker` properties' types have been changed from `int` to `float`.
5. Update of `PInvoke` class in `PInvoke.cs` file. The `DllImport` attribute has been replaced with `LibraryImport` attribute and the methods have been made partial.
6. Update of `Program` class in `Program.cs` file. The options in the `Question.Ask` method in `CacheStuff` and `GenerateCache` methods have been changed. The switch-case statements in `GenerateCache` method have been updated to match the new options.
7. Removal of `Print.PrintCache` method call from the `GenerateCache` method in `Program` class in `Program.cs` file.

+++ JustDanceEditor/JD2Next/KaraokeTape.cs
@@ -25,6 +25,6 @@ public class KaraokeClip
     public int ContentType { get; set; }
     public int StartTimeTolerance { get; set; }
     public int EndTimeTolerance { get; set; }
-    public int SemitoneTolerance { get; set; }
+    public float SemitoneTolerance { get; set; }
 }
